move logic out of useEvents
